Refactored code to move helpers to utils - currently broken
